[fe#28] rename: card -> product card 로 변경
